Fix ruff complaints and consolidate experiment script
- experiments/experiment_nyu.py: replace with improved version from
  nyu_v100.py (timeout logic, MRR, JSON output, Embedding matcher);
  fix E402 noqa for intentional post-warning imports, PLC0415 noqa
  for optional sentence-transformers import; ruff --fix for import sort
- experiments/nyu_v100.py: removed (consolidated into experiment_nyu.py)
